detections/requirements.txt
torch==2.0.1
torchvision==0.15.2
yolov5==7.0.13
ultralytics>=8.2.0
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
scipy==1.11.2
pyyaml==6.0
requests>=2.32.2
urllib3>=2.6.0
Flask==3.0.3
transformers>=4.44.0
